INITIAL COMMIT
The app is already rather advanced, so let's skim over the important details:
We use an MVVM approach with a preference for fragments, although the CameraX is
embedded into its own activity for easier permission handling.
There is already a bunch of libraries in use, so let's try to not use more.
For the ISBN query, we rely on the OpenLibrary-API.